+++ src/utils/rectangle.ts
@@ -1,57 +1,14 @@
 import robot from 'robotjs';
+import {
+  moveDown, moveLeft, moveRight, moveUp,
+} from './line';
 
-export const moveRight = (x: number, y: number, coord: number)
-: Promise<number> => new Promise((resolve) => {
-  let c = 0;
-  const time = setInterval(() => {
-    c += coord / 10;
-    robot.dragMouse(x + c, y);
+export const rectangle = (x: number, y: number, width: number, height: number) => {
+  robot.mouseToggle('down');
 
-    if (c === +coord) {
-      clearInterval(time);
-      resolve(x);
-    }
-  }, 100);
-});
-
-export const moveDown = (x: number, y: number, coord: number)
-: Promise<number> => new Promise((resolve) => {
-  let c = 0;
-  const time = setInterval(() => {
-    c += coord / 10;
-    robot.dragMouse(x, y + c);
-
-    if (c === +coord) {
-      clearInterval(time);
-      resolve(x);
-    }
-  }, 100);
-});
-
-export const moveLeft = (x: number, y: number, coord: number)
-: Promise<number> => new Promise((resolve) => {
-  let c = 0;
-  const time = setInterval(() => {
-    c += coord / 10;
-    robot.dragMouse(x - c, y);
-
-    if (c === +coord) {
-      clearInterval(time);
-      resolve(x);
-    }
-  }, 100);
-});
-
-export const moveUp = (x: number, y: number, coord: number)
-: Promise<number> => new Promise((resolve) => {
-  let c = 0;
-  const time = setInterval(() => {
-    c += coord / 10;
-    robot.dragMouse(x, y - c);
-
-    if (c === +coord) {
-      clearInterval(time);
-      resolve(x);
-    }
-  }, 100);
-});
+  moveRight(x, y, width)
+    .then(() => moveDown(x + width, y, height))
+    .then(() => moveLeft(x + width, y + height, width))
+    .then(() => moveUp(x, y + height, height))
+    .then(() => robot.mouseToggle('up'));
+};
