+++ tsconfig.json
@@ -6,9 +6,6 @@
         "es2020"
       ] /* Specify a set of bundled library declaration files that describe the target runtime environment. */,
   
-      // "baseUrl": "./",
-      // "typeRoots": ["./typings"],
-      // "rootDir": "./src" /* Specify the root folder within your source files. */,
       "outDir": "./dist" /* Specify an output folder for all emitted files. */,
       "moduleResolution": "node",
       "removeComments": true /* Disable emitting comments. */,
